Refs #22875: Apply Miguel's review
