+++ include/fastdds/rtps/builtin/data/SubscriptionBuiltinTopicData.hpp
@@ -47,27 +47,7 @@ struct SubscriptionBuiltinTopicData
             const size_t max_unicast_locators,
             const size_t max_multicast_locators,
             const VariableLengthDataLimits& data_limits,
-            const fastdds::rtps::ContentFilterProperty::AllocationConfiguration& content_filter_limits)
-        : content_filter(content_filter_limits)
-        , remote_locators(max_unicast_locators, max_multicast_locators)
-    {
-        user_data.set_max_size(data_limits.max_user_data);
-        partition.set_max_size(static_cast<uint32_t>(data_limits.max_partitions));
-        data_sharing.set_max_domains(static_cast<uint32_t>(data_limits.max_datasharing_domains));
-    }
-
-    FASTDDS_EXPORTED_API void set_qos(
-            const SubscriptionBuiltinTopicData& qos,
-            bool first_time);
-
-    FASTDDS_EXPORTED_API void set_qos(
-            const dds::ReaderQos& qos,
-            bool first_time);
-
-    FASTDDS_EXPORTED_API bool can_qos_be_updated(
-            const SubscriptionBuiltinTopicData& qos) const;
-
-    FASTDDS_EXPORTED_API void clear();
+            const fastdds::rtps::ContentFilterProperty::AllocationConfiguration& content_filter_limits);
 
     /// Builtin topic Key
     BuiltinTopicKey_t key{{0, 0, 0}};
