+++ src/cpp/CMakeLists.txt
@@ -117,6 +117,7 @@ set(${PROJECT_NAME}_source_files
     rtps/attributes/ServerAttributes.cpp
     rtps/attributes/ThreadSettings.cpp
     rtps/builtin/BuiltinProtocols.cpp
+    rtps/builtin/data/ParticipantBuiltinTopicData.cpp
     rtps/builtin/data/ParticipantProxyData.cpp
     rtps/builtin/data/PublicationBuiltinTopicData.cpp
     rtps/builtin/data/SubscriptionBuiltinTopicData.cpp
