+++ src/cpp/rtps/builtin/data/WriterProxyData.hpp
@@ -363,6 +363,28 @@ public:
             NetworkFactory& network,
             const ParticipantProxyData& participant_data);
 
+    /**
+     * Set qos parameters
+     * (only certain qos from PublicationBuiltinTopicData will be set).
+     *
+     * @param qos Const reference to the SubscriptionBuiltinTopicData object.
+     * @param first_time Boolean indicating whether is the first time (true) or not (false).
+     */
+    void set_qos(
+            const PublicationBuiltinTopicData& qos,
+            bool first_time);
+
+    /**
+     * Set qos parameters from a ReaderQos structure.
+     * (only certain qos from PublicationBuiltinTopicData will be set).
+     *
+     * @param qos Const reference to the ReaderQos object.
+     * @param first_time Boolean indicating whether is the first time (true) or not (false).
+     */
+    void set_qos(
+            const dds::WriterQos& qos,
+            bool first_time);
+
 private:
 
     /**
@@ -370,6 +392,15 @@ private:
      */
     void init(
             const VariableLengthDataLimits& data_limits);
+
+    /**
+     * Checks whether the QoS can be updated with the provided QoS.
+     *
+     * @param qos The QoS to check.
+     * @return true if the QoS can be updated, false otherwise.
+     */
+    bool can_qos_be_updated(
+            const PublicationBuiltinTopicData& qos) const;
 
     //!Network configuration
     NetworkConfigSet_t m_network_configuration;
